requirements-core.txt
numpy>=1.26
scipy>=1.11
pandas>=2.1
matplotlib>=3.8
datasets>=2.20
huggingface_hub>=0.24
pyarrow>=15.0
sentence-transformers>=3.0
mpi4py>=3.1
umap-learn>=0.5.6
